backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic_settings
psycopg2-binary
pydantic
python-jose
passlib
python-multipart
python-dotenv
email-validator
bcrypt